Address luciano's feedback and some extra refactor
