+++ Gemfile
@@ -1,5 +1,7 @@
 source 'http://rubygems.org'
 
+ruby '2.6.6'
+
 gem 'sinatra', '~> 2.0'
 gem 'jwt', '~> 2.1'
 gem 'octokit', '~> 4.0'
