+++ models/article_handler.rb
